fns-tax-mcp/pyproject.toml
[project]
name = "fns-tax-mcp"
version = "1.0.0"
description = "MCP-сервер для генерации и отправки налоговых деклараций через API ФНС России"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.24.0",
    "fastapi>=0.104.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["tools"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["test"]

